+++ src/GraphQL/GraphQL.php
@@ -8,6 +8,7 @@ use Closure;
 use Exception;
 use GraphQL\Error\Debug;
 use GraphQL\Executor\Executor;
+use GraphQL\Executor\Promise\Adapter\SyncPromiseAdapter;
 use GraphQL\Executor\Promise\Promise;
 use GraphQL\Executor\Promise\PromiseAdapter;
 use GraphQL\GraphQL;
@@ -19,16 +20,21 @@ use GraphQL\Type\Schema;
 use Laminas\Diactoros\Response\JsonResponse;
 use Psr\Http\Message\ServerRequestInterface;
 use Ratchet\Server\IoServer;
+use Siler\Arr;
 use Siler\Container;
 use Siler\Diactoros;
+use Siler\GraphQL\Request as GraphQLRequest;
 use Siler\Http\Request;
 use Siler\Http\Response;
 use UnexpectedValueException;
 use WebSocket\BadOpcodeException;
 use WebSocket\Client;
 use function Siler\array_get;
+use function Siler\array_get_arr;
+use function Siler\array_get_str;
 use function Siler\Encoder\Json\decode;
 use function Siler\Encoder\Json\encode;
+use function Siler\GraphQL\request as graphql_request;
 use function Siler\Ratchet\graphql_subscriptions;
 
 // Protocol messages.
@@ -82,7 +88,7 @@ function debugging(): int
  */
 function init(Schema $schema, $rootValue = null, $context = null, string $input = 'php://input'): void
 {
-    $result = execute($schema, input($input), $rootValue, $context);
+    $result = execute($schema, graphql_request($input)->toArray(), $rootValue, $context);
     Response\json($result);
 }
 
@@ -92,6 +98,7 @@ function init(Schema $schema, $rootValue = null, $context = null, string $input 
  * @param string $input
  * @return array<string, mixed>
  * @throws Exception
+ * @deprecated Use request() instead.
  */
 function input(string $input = 'php://input'): array
 {
@@ -113,35 +120,76 @@ function input(string $input = 'php://input'): array
 }
 
 /**
+ * Creates a GraphQL Request based on the current Request (handles the multipart/form-data case on uploads).
+ *
+ * @param string $input
+ * @return GraphQLRequest
+ */
+function request(string $input = 'php://input'): GraphQLRequest
+{
+    /** @var array<string, mixed> $body */
+    $body = Request\body_parse($input);
+
+    if (Request\is_multipart()) {
+        /** @var array<string, mixed> $ops */
+        $ops = decode(array_get_str($body, 'operations'));
+        /** @var array<int, string[]> $map */
+        $map = decode(array_get_str($body, 'map'));
+        $query = array_get_str($ops, 'query');
+        $vars = array_get_arr($ops, 'variables');
+        $op_name = array_get_str($ops, 'operationName');
+
+        foreach ($map as $file_key => $ops_paths) {
+            $file = Request\file($file_key);
+
+            foreach ($ops_paths as $ops_path) {
+                Arr\set($ops, $ops_path, $file);
+            }
+        }
+
+        return new GraphQLRequest($query, $vars, $op_name);
+    }
+
+    $query = array_get_str($body, 'query');
+    $vars = array_get_arr($body, 'variables', []);
+    $op_name = array_get_str($body, 'operationName', '');
+
+    return new GraphQLRequest($query, $vars, $op_name);
+}
+
+/**
  * Executes a GraphQL query over a schema.
  *
+ * @template RootValue
+ * @template Context
  * @param Schema $schema The application root Schema
  * @param array<string, mixed> $input Incoming query, operation and variables
- * @param mixed $rootValue Some optional GraphQL root value
- * @param mixed $context Some optional GraphQL context
+ * @param mixed $rootValue
+ * @psalm-param RootValue|null $rootValue
+ * @param mixed $context
+ * @psalm-param Context|null $context
  *
  * @return array
  */
 function execute(Schema $schema, array $input, $rootValue = null, $context = null)
 {
-    /** @var string $query */
-    $query = array_get($input, 'query');
-    /** @var string $operation */
-    $operation = array_get($input, 'operationName');
-    /** @var array $variables */
-    $variables = array_get($input, 'variables');
-
-    return GraphQL::executeQuery($schema, $query, $rootValue, $context, $variables, $operation)->toArray(debugging());
+    $promise_adapter = new SyncPromiseAdapter();
+    $promise = promise_execute($promise_adapter, $schema, $input, $rootValue, $context);
+    return $promise_adapter->wait($promise)->toArray(debugging());
 }
 
 /**
  * Same as execute(), but allows passing a custom Promise adapter.
  *
+ * @template RootValue
+ * @template Context
  * @param PromiseAdapter $adapter
  * @param Schema $schema
  * @param array<string, mixed> $input
- * @param null $rootValue
- * @param null $context
+ * @param mixed $rootValue
+ * @psalm-param RootValue|null $rootValue
+ * @param mixed $context
+ * @psalm-param Context|null $context
  *
  * @return Promise
  */
@@ -149,12 +197,12 @@ function promise_execute(PromiseAdapter $adapter, Schema $schema, array $input, 
 {
     /** @var string $query */
     $query = array_get($input, 'query');
-    /** @var string $operation */
-    $operation = array_get($input, 'operationName');
     /** @var array $variables */
     $variables = array_get($input, 'variables');
+    /** @var string $operation_name */
+    $operation_name = array_get($input, 'operationName');
 
-    return GraphQL::promiseToExecute($adapter, $schema, $query, $rootValue, $context, $variables, $operation);
+    return GraphQL::promiseToExecute($adapter, $schema, $query, $rootValue, $context, $variables, $operation_name);
 }
 
 /**
@@ -211,6 +259,14 @@ function resolvers(array $resolvers)
     $resolver = static function ($source, array $args, $context, ResolveInfo $info) use ($resolvers) {
         /** @var string|null $field_name */
         $field_name = $info->fieldName;
+    Executor::setDefaultFieldResolver(
+    /**
+     * @param mixed $source
+     * @param mixed $context
+     */
+        static function ($source, array $args, $context, ResolveInfo $info) use ($resolvers) {
+            /** @var string|null $field_name */
+            $field_name = $info->fieldName;
 
         if ($field_name === null) {
             throw new UnexpectedValueException('Could not get fieldName from ResolveInfo');
@@ -305,6 +361,7 @@ function subscriptions_manager(
  * @param array $context
  * @return IoServer
  * @deprecated Returns a new websocket server bootstrapped for GraphQL.
+ * @deprecated Returns a new websocket server bootstrapped for GraphQL
  * @noinspection PhpTooManyParametersInspection
  */
 function subscriptions(
@@ -325,7 +382,7 @@ function subscriptions(
  * @param string $url
  * @return void
  */
-function subscriptions_at(string $url)
+function subscriptions_at(string $url): void
 {
     Container\set('graphql_subscriptions_endpoint', $url);
 }
@@ -349,7 +406,13 @@ function publish(string $subscriptionName, $payload = null): void
     /** @var string $ws_endpoint */
     $ws_endpoint = Container\get('graphql_subscriptions_endpoint');
 
-    $client = new Client($ws_endpoint);
+    $opts = [
+        'headers' => [
+            'Sec-WebSocket-Protocol' => WEBSOCKET_SUB_PROTOCOL
+        ]
+    ];
+
+    $client = new Client($ws_endpoint, $opts);
     $client->send(encode($message));
 }