+++ src/facades.php
@@ -107,6 +107,7 @@ const debug = '\Siler\GraphQL\debug';
 const debugging = '\Siler\GraphQL\debugging';
 const init = '\Siler\GraphQL\init';
 const input = '\Siler\GraphQL\input';
+const request = '\Siler\GraphQL\request';
 const execute = '\Siler\GraphQL\execute';
 const promise_execute = '\Siler\GraphQL\promise_execute';
 const psr7 = '\Siler\GraphQL\psr7';
@@ -139,6 +140,10 @@ namespace Siler\Http\Request;
 const raw = '\Siler\Http\Request\raw';
 const params = '\Siler\Http\Request\params';
 const json = '\Siler\Http\Request\json';
+const body_parse = '\Siler\Http\Request\body_parse';
+const is_json = '\Siler\Http\Request\is_json';
+const is_multipart = '\Siler\Http\Request\is_multipart';
+const content_type = '\Siler\Http\Request\content_type';
 const headers = '\Siler\Http\Request\headers';
 const header = '\Siler\Http\Request\header';
 const get = '\Siler\Http\Request\get';
@@ -190,9 +195,22 @@ const alert = '\Siler\Monolog\alert';
 const emergency = '\Siler\Monolog\emergency';
 
 
+namespace Siler\Arr;
+const set = '\Siler\Arr\set';
+
+
+namespace Siler\Klass;
+const unqualified_name = '\Siler\Klass\unqualified_name';
+
+
 namespace Siler\Str;
 const slugify = '\Siler\Str\slugify';
 const lines = '\Siler\Str\lines';
+const starts_with = '\Siler\Str\starts_with';
+const ends_with = '\Siler\Str\ends_with';
+const contains = '\Siler\Str\contains';
+const snake_case = '\Siler\Str\snake_case';
+const camel_case = '\Siler\Str\camel_case';
 
 
 namespace Siler\Tuple;
@@ -238,6 +256,7 @@ const array_get_str = '\Siler\array_get_str';
 const array_get_int = '\Siler\array_get_int';
 const array_get_float = '\Siler\array_get_float';
 const array_get_bool = '\Siler\array_get_bool';
+const array_get_arr = '\Siler\array_get_arr';
 const require_fn = '\Siler\require_fn';