+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.7.3
+- GraphQL Uploads!
+- Fix #276 - GraphQL WebSocket client sub-protocol
+- PHPUnit 9
+- Better error message for `array_get`'s
+- `Siler\array_get_arr` type-safe array getter for arrays
+- `Str\starts_with`, `Str\ends_with` and `Str\contains`
+- Introducing Klass module with `Klass\unqualified_name`
+- `Str\snake_case` and `Str\camel_case` case converters
+
 # 1.7.2
 - Experimental support for gRPC servers
 - Support middleware-like pipelines in Swoole with `Swoole\middleware`